requirements.txt
discord.py
apscheduler
pandas
gspread
oauth2client
